begin sharing flow for Content types
