+++ packages/commuter-server/src/content-providers/local/fs.js
@@ -7,49 +7,12 @@
 const fs = require("fs");
 const path = require("path");
 
-export type DirectoryContent = {
-  type: "directory",
-  mimetype: null,
-  content: null | Array<Content>, // Technically content-free content ;)
-
-  name: string,
-  path: string,
-
-  created: Date,
-  last_modified: Date,
-  writable: boolean,
-  format: "json"
-};
-
-export type NotebookContent = {
-  type: "notebook",
-  mimetype: null,
-  content: null | Object,
-
-  name: string,
-  path: string,
-
-  created: Date,
-  last_modified: Date,
-  writable: boolean,
-  format: "json"
-};
-
-export type FileContent = {
-  type: "file",
-  mimetype: null | string,
-  content: null | string,
-
-  name: string,
-  path: string,
-
-  created: Date,
-  last_modified: Date,
-  writable: boolean,
-  format: null | "text" | "base64"
-};
-
-export type Content = DirectoryContent | FileContent | NotebookContent;
+import type {
+  Content,
+  DirectoryContent,
+  FileContent,
+  NotebookContent
+} from "../base";
 
 export type DiskProviderOptions = {
   baseDirectory: string
